requirements.txt
google-generativeai
rich
click
